+++ Backend/.env
@@ -1,0 +1,2 @@
+JWT_KEY = OUR_ONDC_PROJECT
+MONGO_URI = mongodb://localhost:27017/zenNexify
